fixed small bugs in store

+++ src/main/java/controller/GameControllers/MarketController.java
@@ -39,7 +39,7 @@ public class MarketController {
         StoreMaterial storeMaterial = StoreMaterial.getInstance(materialName);
         Material material = Material.getMaterial(materialName);
         if (storeMaterial == null) return MarketMessage.INVALID_MATERIAL;
-        int gold = currentPlayer.getGovernance().getGold();
+        int gold = (int) currentPlayer.getGovernance().getGold();
         if (gold < storeMaterial.getPrice()*amount) return MarketMessage.NOT_ENOUGH_GOLD;
         if (menu.confirm(confirmBuyMessage(materialName, amount,
                 storeMaterial.getPrice()*amount))){
